added new resist_poison_gas spell
Added it to the 4th mage and 4th priest books.
Only mages and priests can use resist_poison_gas, to reduce poison gas damage by dividing it by 1.5. this reduces the maximum breath attack from 347 to 231 of level 31 characters who can learn this spell.
This spell gives poison gas resistance permanently.
This new spell is meant to offset the late threat of AMHD gas attacks insta-killing characters with a small amount of hp.
This spell is affects all race/class combinations that have a starting hitpoints of 11 or lower.  Human, Half-Orc, Half-Troll priests get to cast this spell even though they have starting hitpoints of 12 or more -- so their game gets a little bit easier, even though they don't really need the poison resistance because their hp at the end of the game is generally high enough.
This also has the side effect of changing the lettering of the final 2 spells and prayers.  if these were early spells it would be more of a problem.